requirements.txt
torch
torchvision
scikit-learn
Pillow
